refactor: move around scripts, reuse code
