+++ hrm-domain/speech-to-text/proxy-service/src/commands/cli.ts
@@ -14,13 +14,22 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-export const command = 'limina <command>';
-export const desc = 'endpoints for limina';
-export const builder = function (yargs) {
-  return yargs.commandDir('limina', {
-    exclude: /^(index|_)/, // Exclude files starting with 'index' or '_'
-    extensions: ['ts'],
-  });
-  // .commandDir('common_cmds'); add more
+// eslint-disable-next-line import/no-extraneous-dependencies
+import yargs from 'yargs';
+
+const main = () => {
+  console.info('Commands CLI');
+  // eslint-disable-next-line @typescript-eslint/no-unused-expressions
+  yargs
+    .commandDir('.', {
+      exclude: /^(index|_)/, // Exclude files starting with 'index' or '_'
+      extensions: ['js', 'ts'],
+    })
+    .scriptName('commands-cli')
+    .demandCommand(1, 'Please provide a valid command')
+    .version(false)
+    .wrap(120)
+    .help().argv;
 };
-// export const handler = function (argv) {};
+
+main();
