+++ hrm-domain/speech-to-text/proxy-service/src/commands/limina/healthcheck.ts
