+++ hrm-domain/speech-to-text/proxy-service/src/commands/pyannote/healthcheck.ts
